+++ unsloth/kernels/moe/benchmark/benchmark_fused_moe.py
@@ -56,7 +56,7 @@ def run_benchmark_forward(
     hidden_size = config.hidden_size
 
     X = torch.randn(
-        bs, seqlen, hidden_size, dtype=dtype, device=device, requires_grad=True
+        bs, seqlen, hidden_size, dtype = dtype, device = device, requires_grad = True
     )
 
     # Forward
@@ -89,7 +89,7 @@ def run_benchmark_backward(
     config: AutoConfig,
     seqlen: int,
     dtype: torch.dtype,
-    bs=1,
+    bs = 1,
 ):
     torch.manual_seed(
         SEED
@@ -98,7 +98,7 @@ def run_benchmark_backward(
     hidden_size = config.hidden_size
 
     X = torch.randn(
-        bs, seqlen, hidden_size, dtype=dtype, device=device, requires_grad=True
+        bs, seqlen, hidden_size, dtype = dtype, device = device, requires_grad = True
     )
     X_test = X.detach().clone().requires_grad_(True)
 
@@ -114,14 +114,14 @@ def run_benchmark_backward(
 
     # Bench
     grad_output = torch.randn_like(output)
-    bench_backward_ref = lambda: output.backward(grad_output, retain_graph=True)  # noqa: E731
-    bench_backward_fused = lambda: test_output.backward(grad_output, retain_graph=True)  # noqa: E731
+    bench_backward_ref = lambda: output.backward(grad_output, retain_graph = True)  # noqa: E731
+    bench_backward_fused = lambda: test_output.backward(grad_output, retain_graph = True)  # noqa: E731
 
     ref_backward_time = do_bench(
-        bench_backward_ref, grad_to_none=[X, *ref_model.parameters()]
+        bench_backward_ref, grad_to_none = [X, *ref_model.parameters()]
     )
     fused_backward_time = do_bench(
-        bench_backward_fused, grad_to_none=[X_test, *tt_model.parameters()]
+        bench_backward_fused, grad_to_none = [X_test, *tt_model.parameters()]
     )
     print(
         f"Backward: ref {ref_backward_time:.4f}, fused {fused_backward_time:.4f}, speedup {ref_backward_time / fused_backward_time:.1f}x"
@@ -138,10 +138,10 @@ def setup_model(
     kernel_config_fwd,
     kernel_config_bwd_dW,
     kernel_config_bwd_dX,
-    dX_only=False,
-    dW_only=False,
-    overlap_router_shared=False,
-    device="cuda",
+    dX_only = False,
+    dW_only = False,
+    overlap_router_shared = False,
+    device = "cuda",
 ):
     if isinstance(config, Qwen3MoeConfig):
         ref_model = Qwen3MoeSparseMoeBlock(config).to(device, dtype)
@@ -149,29 +149,29 @@ def setup_model(
         # Triton kernel grouped gemm version of MoE Block -- this is what we're testing
         tt_model = Qwen3MoeFusedGroupedGEMMBlock.from_hf(
             ref_model,
-            permute_x=permute_x,
-            permute_y=permute_y,
-            autotune=autotune,
-            kernel_config_fwd=kernel_config_fwd,
-            kernel_config_bwd_dW=kernel_config_bwd_dW,
-            kernel_config_bwd_dX=kernel_config_bwd_dX,
-            dX_only=dX_only,
-            dW_only=dW_only,
+            permute_x = permute_x,
+            permute_y = permute_y,
+            autotune = autotune,
+            kernel_config_fwd = kernel_config_fwd,
+            kernel_config_bwd_dW = kernel_config_bwd_dW,
+            kernel_config_bwd_dX = kernel_config_bwd_dX,
+            dX_only = dX_only,
+            dW_only = dW_only,
         ).to(device, dtype)
 
     elif isinstance(config, Llama4TextConfig):
         ref_model = Llama4TextMoe(config).to(device, dtype)
         tt_model = Llama4TritonTextMoe(
             config,
-            overlap_router_shared=overlap_router_shared,
-            permute_x=permute_x,
-            permute_y=permute_y,
-            autotune=autotune,
-            kernel_config_fwd=kernel_config_fwd,
-            kernel_config_bwd_dW=kernel_config_bwd_dW,
-            kernel_config_bwd_dX=kernel_config_bwd_dX,
-            dX_only=dX_only,
-            dW_only=dW_only,
+            overlap_router_shared = overlap_router_shared,
+            permute_x = permute_x,
+            permute_y = permute_y,
+            autotune = autotune,
+            kernel_config_fwd = kernel_config_fwd,
+            kernel_config_bwd_dW = kernel_config_bwd_dW,
+            kernel_config_bwd_dX = kernel_config_bwd_dX,
+            dX_only = dX_only,
+            dW_only = dW_only,
         ).to(device, dtype)
 
     else:
@@ -205,31 +205,31 @@ def run_benchmark(
 
     ref_model, tt_model = setup_model(
         model_config,
-        dtype=dtype,
-        permute_x=permute_x,
-        permute_y=permute_y,
-        autotune=autotune,
-        kernel_config_fwd=kernel_config_fwd,
-        kernel_config_bwd_dW=kernel_config_bwd_dW,
-        kernel_config_bwd_dX=kernel_config_bwd_dX,
-        dX_only=dX_only,
-        dW_only=dW_only,
-        overlap_router_shared=overlap_router_shared,
+        dtype = dtype,
+        permute_x = permute_x,
+        permute_y = permute_y,
+        autotune = autotune,
+        kernel_config_fwd = kernel_config_fwd,
+        kernel_config_bwd_dW = kernel_config_bwd_dW,
+        kernel_config_bwd_dX = kernel_config_bwd_dX,
+        dX_only = dX_only,
+        dW_only = dW_only,
+        overlap_router_shared = overlap_router_shared,
     )
 
     if mode == "forward":
         ref_time, fused_time = run_benchmark_forward(
             ref_model,
             tt_model,
-            config=model_config,
-            seqlen=seqlen,
-            dtype=dtype,
-            autotune=autotune,
-            kernel_config_fwd=kernel_config_fwd,
+            config = model_config,
+            seqlen = seqlen,
+            dtype = dtype,
+            autotune = autotune,
+            kernel_config_fwd = kernel_config_fwd,
         )
     else:
         ref_time, fused_time = run_benchmark_backward(
-            ref_model, tt_model, config=model_config, seqlen=seqlen, dtype=dtype
+            ref_model, tt_model, config = model_config, seqlen = seqlen, dtype = dtype
         )
 
     if autotune:
@@ -251,60 +251,60 @@ def run_benchmark(
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument("--results_dir", type=str, default="benchmark_results")
-    parser.add_argument("--model", type=str, choices=["llama4", "qwen3"], required=True)
-    parser.add_argument("--seqlen", type=int, default=1024)
+    parser.add_argument("--results_dir", type = str, default = "benchmark_results")
+    parser.add_argument("--model", type = str, choices = ["llama4", "qwen3"], required = True)
+    parser.add_argument("--seqlen", type = int, default = 1024)
     parser.add_argument(
-        "--dtype", type=str, choices=["bfloat16", "float16"], default="bfloat16"
+        "--dtype", type = str, choices = ["bfloat16", "float16"], default = "bfloat16"
     )
-    parser.add_argument("--permute_x", action="store_true")
-    parser.add_argument("--permute_y", action="store_true")
-    parser.add_argument("--autotune", action="store_true")
-    parser.add_argument("--overlap_router_shared", action="store_true")
+    parser.add_argument("--permute_x", action = "store_true")
+    parser.add_argument("--permute_y", action = "store_true")
+    parser.add_argument("--autotune", action = "store_true")
+    parser.add_argument("--overlap_router_shared", action = "store_true")
     parser.add_argument(
         "--BLOCK_SIZE_M",
-        nargs=2,
-        type=int,
-        default=[DEFAULT_M_BLOCK_SIZES[0], DEFAULT_M_BLOCK_SIZES[-1]],
+        nargs = 2,
+        type = int,
+        default = [DEFAULT_M_BLOCK_SIZES[0], DEFAULT_M_BLOCK_SIZES[-1]],
     )
     parser.add_argument(
         "--BLOCK_SIZE_N",
-        nargs=2,
-        type=int,
-        default=[DEFAULT_N_BLOCK_SIZES[0], DEFAULT_N_BLOCK_SIZES[-1]],
+        nargs = 2,
+        type = int,
+        default = [DEFAULT_N_BLOCK_SIZES[0], DEFAULT_N_BLOCK_SIZES[-1]],
     )
     parser.add_argument(
         "--BLOCK_SIZE_K",
-        nargs=2,
-        type=int,
-        default=[DEFAULT_K_BLOCK_SIZES[0], DEFAULT_K_BLOCK_SIZES[-1]],
+        nargs = 2,
+        type = int,
+        default = [DEFAULT_K_BLOCK_SIZES[0], DEFAULT_K_BLOCK_SIZES[-1]],
     )
     parser.add_argument(
         "--num_warps",
-        nargs=2,
-        type=int,
-        default=[DEFAULT_NUM_WARPS[0], DEFAULT_NUM_WARPS[-1]],
+        nargs = 2,
+        type = int,
+        default = [DEFAULT_NUM_WARPS[0], DEFAULT_NUM_WARPS[-1]],
     )
     parser.add_argument(
         "--num_stages",
-        nargs=2,
-        type=int,
-        default=[DEFAULT_NUM_STAGES[0], DEFAULT_NUM_STAGES[-1]],
+        nargs = 2,
+        type = int,
+        default = [DEFAULT_NUM_STAGES[0], DEFAULT_NUM_STAGES[-1]],
     )
     parser.add_argument(
-        "--use_tma_load_w", action="store_true"
+        "--use_tma_load_w", action = "store_true"
     )  # No need to specify, will automatically parametrize these for each kernel config
     parser.add_argument(
-        "--use_tma_load_x", action="store_true"
+        "--use_tma_load_x", action = "store_true"
     )  # No need to specify, will automatically parametrize these for each kernel config
     parser.add_argument(
-        "--use_tma_load_dy", action="store_true"
+        "--use_tma_load_dy", action = "store_true"
     )  # No need to specify, will automatically parametrize these for each kernel config
     parser.add_argument(
         "--mode",
-        type=str,
-        choices=["forward", "backward", "dW", "dX"],
-        default="forward",
+        type = str,
+        choices = ["forward", "backward", "dW", "dX"],
+        default = "forward",
     )
     args = parser.parse_args()
     args.dtype = getattr(torch, args.dtype)
@@ -324,13 +324,13 @@ if __name__ == "__main__":
         ref_time, fused_time = run_benchmark(
             args.mode,
             model_config,
-            seqlen=args.seqlen,
-            dtype=args.dtype,
-            permute_x=args.permute_x,
-            permute_y=args.permute_y,
-            autotune=args.autotune,
-            overlap_router_shared=args.overlap_router_shared,
-            results_dir=args.results_dir,
+            seqlen = args.seqlen,
+            dtype = args.dtype,
+            permute_x = args.permute_x,
+            permute_y = args.permute_y,
+            autotune = args.autotune,
+            overlap_router_shared = args.overlap_router_shared,
+            results_dir = args.results_dir,
         )
         end_time = time.time()
         print(f"Total time: {end_time - start_time:.4f} seconds")
@@ -343,13 +343,13 @@ if __name__ == "__main__":
         kernel_configs = create_kernel_configs(args, args.permute_x, args.permute_y)
         print(f"Running {len(kernel_configs)} kernel configs")
         default_kernel_config_fwd = KernelConfigForward(
-            permute_x=args.permute_x, permute_y=args.permute_y
+            permute_x = args.permute_x, permute_y = args.permute_y
         )
         default_kernel_config_bwd_dW = KernelConfigBackward_dW(
-            permute_x=args.permute_x, permute_y=args.permute_y
+            permute_x = args.permute_x, permute_y = args.permute_y
         )
         default_kernel_config_bwd_dX = KernelConfigBackward_dX(
-            permute_x=args.permute_x, permute_y=args.permute_y
+            permute_x = args.permute_x, permute_y = args.permute_y
         )
         results = []
         for kernel_config in kernel_configs:
@@ -374,21 +374,21 @@ if __name__ == "__main__":
             ref_time, fused_time = run_benchmark(
                 args.mode,
                 model_config,
-                seqlen=args.seqlen,
-                dtype=args.dtype,
-                permute_x=kernel_config.permute_x,
-                permute_y=kernel_config.permute_y,
-                autotune=False,
-                kernel_config_fwd=kernel_config_fwd,
-                kernel_config_bwd_dW=kernel_config_bwd_dW,
-                kernel_config_bwd_dX=kernel_config_bwd_dX,
+                seqlen = args.seqlen,
+                dtype = args.dtype,
+                permute_x = kernel_config.permute_x,
+                permute_y = kernel_config.permute_y,
+                autotune = False,
+                kernel_config_fwd = kernel_config_fwd,
+                kernel_config_bwd_dW = kernel_config_bwd_dW,
+                kernel_config_bwd_dX = kernel_config_bwd_dX,
             )
             results.append(
                 KernelResult(
-                    torch_time=ref_time,
-                    triton_time=fused_time,
-                    speedup=ref_time / fused_time,
-                    kernel_config=kernel_config,
+                    torch_time = ref_time,
+                    triton_time = fused_time,
+                    speedup = ref_time / fused_time,
+                    kernel_config = kernel_config,
                 )
             )
         df = post_process_results(
